requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ZMQ communication (install from git)
# zmqhelpers - install manually from https://github.com/kshalm/zmqhelpers
pyzmq==25.1.0
requests==2.31.0
loguru==0.7.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Redis for data monitoring
redis==5.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development dependencies
black==23.11.0
flake8==6.1.0